flask-auth0/pyproject.toml
[tool.poetry]
name = "auth0-flask"
version = "0.1.0"
description = ""
authors = ["Attila Domokos <adomokos@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
authlib = "^0.14.3"
six = "^1.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
run-server = "server:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
